+++ src/test/kotlin/no/nav/helse/tildeling/TildelingDatabaseFactory.kt
@@ -32,7 +32,7 @@ fun DataSource.opprettVedtak() = sessionOf(this, returnGeneratedKey = true).use 
         fom = LocalDate.now(),
         tom = LocalDate.now(),
         personRef = person!!.toInt(),
-        arbeidsgiverRef = arbeidsgiver!!.toInt(),
+        arbeidsgiverRef = arbeidsgiver!!,
         speilSnapshotRef = speilSnapshot
     )
 }
